+++ test/sample/foo.js
@@ -1,4 +1,4 @@
-var gobble = require( '../../tmp' ).default,
+var gobble = require( '../..' ),
 	path = require( 'path' );
 
 module.exports = gobble( path.join( __dirname, 'foo' ) );